+++ src/PF/PF_utils.cpp
@@ -3,8 +3,7 @@
 
 inline covarmat set_bw_fw_particle_combiner_Q
   (const PF_data &data, const covarmat &Q_trans){
-  arma::mat Q_inv_arg = data.err_state_inv->map(
-    Q_trans.inv(), both, trans).sv;
+  arma::mat Q_inv_arg = data.err_state->map(Q_trans.inv()).sv;
   Q_inv_arg += data.state_trans->map(Q_inv_arg, both, trans).sv;
   return covarmat(Q_inv_arg.i());
 }
@@ -21,13 +20,17 @@ arma::vec bw_fw_particle_combiner::operator()
 
 arma::vec bw_fw_particle_combiner::operator()
   (const arma::vec &fw_p, const arma::vec &bw_p, const bool do_transform) const {
-  /* compute part of the mean from the forward particle */
+  /* compute part of the mean from the forward particle
+   * R Q^{-1}R^\top Fx_{t - 1}
+   * TODO: issues w/ higher orders currently... */
   arma::vec mu = data.state_trans_err->map(fw_p).sv;
   mu = solve_w_precomputed_chol(Q_trans.chol(), mu);
-  mu = data.state_trans_err->map(mu, trans).sv;
+  mu = data.err_state->map(mu).sv;
 
   /* add part of the mean from the backward particle */
   {
+    /* F^\top R Q^{-1} R^\top x_{t + 1}
+     * TODO: issues w/ higher orders currently... */
     arma::vec bw_term = data.err_state_inv->map(bw_p).sv;
     bw_term = solve_w_precomputed_chol(Q_trans.chol(), bw_term);
     mu += data.state_trans_err->map(bw_term, trans).sv;